Начата разработка модуля Event. Сделана публичная часть и почти админская.

+++ main-service/src/main/java/ru/yandex/practicum/user/dto/UserShortDto.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserShortDto {
-    private Long id;
+
+    private Integer id;
     private String name;
 }
